Resolve Imprevisto

+++ src/Controller/ImprevistoController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Imprevisto;
+use App\Entity\Viaje;
 use App\Form\ImprevistoType;
 use App\Repository\ImprevistoRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -49,6 +50,31 @@ class ImprevistoController extends AbstractController
     }
 
     /**
+     * @Route("/{id}/new", name="imprevistoWithViaje_new", methods={"GET","POST"})
+     */
+    public function newWithViaje(Viaje $viaje,Request $request): Response
+    {
+        $imprevisto = new Imprevisto();
+        $form = $this->createForm(ImprevistoType::class, $imprevisto)
+            ->remove('viaje');
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $imprevisto->setViaje($viaje);
+            $entityManager->persist($imprevisto);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('imprevisto_index');
+        }
+
+        return $this->render('imprevisto/new.html.twig', [
+            'imprevisto' => $imprevisto,
+            'form' => $form->createView(),
+        ]);
+    }
+
+    /**
      * @Route("/{id}", name="imprevisto_show", methods={"GET"})
      */
     public function show(Imprevisto $imprevisto): Response
@@ -66,12 +92,16 @@ class ImprevistoController extends AbstractController
         $form = $this->createForm(ImprevistoType::class, $imprevisto);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+        if(!$imprevisto->getState()){
 
-            return $this->redirectToRoute('imprevisto_index');
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+                return $this->redirectToRoute('imprevisto_index');
+            }
+            
+        }else{
+            $this->addFlash('failed','No pueden editarse imprevisto que ya se resolvieron');
         }
-
         return $this->render('imprevisto/edit.html.twig', [
             'imprevisto' => $imprevisto,
             'form' => $form->createView(),
@@ -83,12 +113,35 @@ class ImprevistoController extends AbstractController
      */
     public function delete(Request $request, Imprevisto $imprevisto): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$imprevisto->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($imprevisto);
-            $entityManager->flush();
+        
+        if(!$imprevisto->getState()){
+
+            if ($this->isCsrfTokenValid('delete'.$imprevisto->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($imprevisto);
+                $entityManager->flush();
+            }
+
+        }else{
+            $this->addFlash('failed','No pueden eliminsar imprevisto que ya se resolvieron');
         }
+        
 
         return $this->redirectToRoute('imprevisto_index');
+    }
+
+    /**
+     * @Route("/{id}/resolve", name="imprevisto_resolve", methods={"POST"})
+     */
+    public function resolve(Imprevisto $imprevisto, ImprevistoRepository $imprevistoRepository): Response
+    {
+        $entityManager = $this->getDoctrine()->getManager();
+        $imprevisto->resolver();
+        $entityManager->persist($imprevisto);
+        $entityManager->flush();
+
+        return $this->render('imprevisto/index.html.twig', [
+            'imprevistos' => $imprevistoRepository->findAll(),
+        ]);
     }
 }
